+++ symja_android_library/matheclipse-core/src/main/java/org/matheclipse/core/convert/AST2Expr.java
@@ -276,7 +276,7 @@ public class AST2Expr {
 		Initializer.init();
 	}
 
-	private int fPrecision;
+	private long fPrecision;
 
 	private boolean fLowercaseEnabled;
